Desktop/Web-Dev/lab10/hh-back/vacancies/serializers.py
from rest_framework import serializers
from .models import Vacancy
from companies.serializers import CompanyModelSerializer

class VacancySerializer(serializers.Serializer):
    id = serializers.IntegerField(read_only=True)
    name = serializers.CharField(max_length=255)
    description = serializers.CharField()
    salary = serializers.FloatField()
    company = CompanyModelSerializer(read_only=True)
    company_id = serializers.IntegerField(write_only=True)

    def create(self, validated_data):
        return Vacancy.objects.create(**validated_data)

    def update(self, instance, validated_data):
        instance.name = validated_data.get('name', instance.name)
        instance.description = validated_data.get('description', instance.description)
        instance.salary = validated_data.get('salary', instance.salary)
        instance.company_id = validated_data.get('company_id', instance.company_id)
        instance.save()
        return instance