Desktop/Web-Dev/lab10/hh-back/requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
fonttools==4.56.0
inflection==0.5.1
kiwisolver==1.4.8
matplotlib==3.10.1
mpmath==1.3.0
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.0
pdf2image==1.17.0
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pygame==2.5.2
PyJWT==1.7.1
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.2
scipy==1.12.0
six==1.16.0
sqlparse==0.5.3
sympy==1.12
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.29.3
